% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC}
\alias{AIC}
\alias{AIC.ggpd}
\alias{AIC.mgpd}
\title{AIC model selection criterion}
\usage{
AIC(x, ...)

\method{AIC}{ggpd}(x, ...)

\method{AIC}{mgpd}(x, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}
}
\value{
The AIC of a model estimated with \code{extrememix}
}
\description{
Computation of the AIC for an extreme value mixture model
}
