% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{AICc}
\alias{AICc}
\alias{AICc.ggpd}
\alias{AICc.mgpd}
\title{AICc model selection criterion}
\usage{
AICc(x, ...)

\method{AICc}{ggpd}(x, ...)

\method{AICc}{mgpd}(x, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}
}
\value{
The AICc of a model estimated with \code{extrememix}
}
\description{
Computation of the AICc for an extreme value mixture model
}
