% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC}
\alias{BIC}
\alias{BIC.ggpd}
\alias{BIC.mgpd}
\title{BIC model selection criterion}
\usage{
BIC(x, ...)

\method{BIC}{ggpd}(x, ...)

\method{BIC}{mgpd}(x, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}
}
\value{
The BIC of a model estimated with \code{extrememix}
}
\description{
Computation of the BIC for an extreme value mixture model
}
