% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\alias{check_convergence.evmm}
\title{VaR Method}
\usage{
check_convergence(x, ...)

\method{check_convergence}{evmm}(x, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}
}
\description{
Compute and plot VaRs
}
