% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claim.R
\docType{data}
\name{claim}
\alias{claim}
\title{Automobile Claims Data}
\format{
A numeric vector of 4624 insurance claims
}
\source{
Charpentier, A. (Ed.). (2014). Computational actuarial science with R. CRC press.

Goegebeur, Yuri, Armelle Guillou, and Jing Qin. "Extreme value estimation of the conditional risk premium in reinsurance." Insurance: Mathematics and Economics 96 (2021): 68-80.
}
\usage{
data(claim)
}
\description{
Claim amounts of one-year vehicle insurance policies taken out in 2004 or 2005 in Australia
}
\keyword{datasets}
