% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC.R
\name{DIC}
\alias{DIC}
\alias{DIC.ggpd}
\alias{DIC.mgpd}
\title{DIC model selection criterion}
\usage{
DIC(x, ...)

\method{DIC}{ggpd}(x, ...)

\method{DIC}{mgpd}(x, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}
}
\value{
The DIC of a model estimated with \code{extrememix}
}
\description{
Computation of the DIC for an extreme value mixture model
}
