% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES.R
\name{ES}
\alias{ES}
\alias{ES.ggpd}
\alias{ES.mgpd}
\title{Expected Shortfall Method}
\usage{
ES(x, ...)

\method{ES}{ggpd}(x, values = NULL, cred = 0.95, ...)

\method{ES}{mgpd}(x, values = NULL, cred = 0.95, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}

\item{values}{vector of points where to estimate the predictive distribution}

\item{cred}{amplitude of the posterior credibility interval}
}
\description{
Compute and plot Expected Shortfalls
}
