% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmgpd.R
\name{fmgpd}
\alias{fmgpd}
\title{Fit of the GGPD model using an MCMC algorithm.}
\usage{
fmgpd(x, it, k, start = NULL, var = NULL, prior = NULL, thin = 1, burn = 0)
}
\arguments{
\item{x}{A vector of positive observations.}

\item{it}{Number of iterations of the algorithm.}

\item{k}{number of mixture components for the bulk}

\item{start}{A list of starting parameter values.}

\item{var}{A list of starting proposal variance.}

\item{prior}{A list of hyperparameters for the prior distribution.}

\item{thin}{Thinning interval.}

\item{burn}{Burn-in.}
}
\description{
Fit of the GGPD model using an MCMC algorithm.
}
