% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpd.R
\name{ggpd}
\alias{ggpd}
\alias{dggpd}
\alias{pggpd}
\alias{qggpd}
\alias{rggpd}
\title{The GGPD distribution}
\usage{
dggpd(x, xi, sigma, u, mu, eta, log = FALSE)

pggpd(q, xi, sigma, u, mu, eta, lower.tail = TRUE)

qggpd(p, xi, sigma, u, mu, eta, lower.tail = TRUE)

rggpd(N, xi, sigma, u, mu, eta)
}
\arguments{
\item{x}{vector of quantiles.}

\item{xi}{shape parameter of the tail GPD (scalar).}

\item{sigma}{scale parameter of the tail GPD (scalar).}

\item{u}{threshold parameter of the tail GPD (scalar).}

\item{mu}{mean of the gamma bulk (scalar).}

\item{eta}{shape of the gamma bulk (scalar).}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P(X\leq x)} otherwise \eqn{P(X>x)}.}

\item{p}{vector of probabilities.}

\item{N}{number of observations.}
}
\value{
A co-variation matrix of the same size of the covariance matrix of \code{CI}.
}
\description{
Density, distribution function, quantile function and random generation for the GGPD distribution.
}
\examples{
dggpd(3,0.5,2,5,3,3)

}
