% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\alias{logLik.ggpd}
\alias{logLik.mgpd}
\title{Log-likelihood method}
\usage{
logLik(x, ...)

\method{logLik}{ggpd}(x, ...)

\method{logLik}{mgpd}(x, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}
}
\value{
The log-likelihood of a model estimated with \code{extrememix}
}
\description{
Computation of the log-likelihood for an extreme value mixture model
}
