% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgpd.R
\name{mgpd}
\alias{mgpd}
\alias{dmgpd}
\alias{pmgpd}
\alias{qmgpd}
\alias{rmgpd}
\title{The MGPD distribution}
\usage{
dmgpd(x, xi, sigma, u, mu, eta, w, log = FALSE)

pmgpd(q, xi, sigma, u, mu, eta, w, lower.tail = TRUE)

qmgpd(p, xi, sigma, u, mu, eta, w, lower.tail = TRUE)

rmgpd(N, xi, sigma, u, mu, eta, w)
}
\arguments{
\item{x}{vector of quantiles.}

\item{xi}{shape parameter of the tail GPD (scalar).}

\item{sigma}{scale parameter of the tail GPD (scalar).}

\item{u}{threshold parameter of the tail GPD (scalar).}

\item{mu}{means of the gamma mixture components (vector).}

\item{eta}{shapes of the gamma mixture components (vector).}

\item{w}{weights of the gamma mixture components (vector). Must sum to one.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P(X\leq x)} otherwise \eqn{P(X>x)}.}

\item{p}{vector of probabilities.}

\item{N}{number of observations.}
}
\value{
A co-variation matrix of the same size of the covariance matrix of \code{CI}.
}
\description{
Density, distribution function, quantile function and random generation for the MGPD distribution.
}
\examples{
dmgpd(3,0.5,2,5,c(2,3),c(1,2),c(0.3,0.7))

}
