% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ES}
\alias{plot.ES}
\title{Plot ESs}
\usage{
\method{plot}{ES}(x, ylim = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{ES}.}

\item{ylim}{ciao}

\item{...}{additional parameters (compatibility).}
}
\description{
Plot ESs
}
\details{
Info
}
