% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.evmm}
\alias{plot.evmm}
\title{Plot quantiles}
\usage{
\method{plot}{evmm}(x, ...)
}
\arguments{
\item{x}{an object of class \code{evmm}.}

\item{...}{additional parameters (compatibility).}
}
\description{
Plot quantiles
}
\details{
Info
}
