% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.quant}
\alias{plot.quant}
\title{Plot quantiles}
\usage{
\method{plot}{quant}(x, ylim = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{quant}.}

\item{ylim}{ciao}

\item{...}{additional parameters (compatibility).}
}
\description{
Plot quantiles
}
\details{
Info
}
