% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.return_level}
\alias{plot.return_level}
\title{Plot return levels}
\usage{
\method{plot}{return_level}(x, ylim = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{return_level}.}

\item{ylim}{ciao}

\item{...}{additional parameters (compatibility).}
}
\description{
Plot return levels
}
\details{
Info
}
