% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.TVaR}
\alias{plot.TVaR}
\title{Plot TVaRs}
\usage{
\method{plot}{TVaR}(x, ylim = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{TVaR}.}

\item{ylim}{ciao}

\item{...}{additional parameters (compatibility).}
}
\description{
Plot TVaRs
}
\details{
Info
}
