% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.VaR}
\alias{plot.VaR}
\title{Plot VaRs}
\usage{
\method{plot}{VaR}(x, ylim = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{VaR}.}

\item{ylim}{ciao}

\item{...}{additional parameters (compatibility).}
}
\description{
Plot VaRs
}
\details{
Info
}
