% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\alias{pred.ggpd}
\alias{pred.mgpd}
\title{Predictive Distribution}
\usage{
pred(x, ...)

\method{pred}{ggpd}(x, x_axis = seq(0, max(x$data), 0.01), cred = 0.95, ...)

\method{pred}{mgpd}(x, x_axis = seq(0, max(x$data), 0.01), cred = 0.95, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}

\item{x_axis}{vector of points where to estimate the predictive distribution}

\item{cred}{amplitude of the posterior credibility interval}
}
\description{
Plot of the predictive distribution of an extreme value mixture model
}
