% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.ES}
\alias{print.ES}
\title{Print a ES object}
\usage{
\method{print}{ES}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ES}.}

\item{...}{additional parameters (compatibility).}
}
\description{
Print a ES object
}
\details{
Info
}
