% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.ggpd}
\alias{print.ggpd}
\title{Print a ggpd}
\usage{
\method{print}{ggpd}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ggpd}.}

\item{...}{additional parameters (compatibility).}
}
\description{
Print a ggpd
}
\details{
Info
}
