% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.mgpd}
\alias{print.mgpd}
\title{Print a mgpd}
\usage{
\method{print}{mgpd}(x, ...)
}
\arguments{
\item{x}{an object of class \code{mgpd}.}

\item{...}{additional parameters (compatibility).}
}
\description{
Print a mgpd
}
\details{
Info
}
