% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.quantile}
\alias{print.quantile}
\title{Print a quantile object}
\usage{
\method{print}{quantile}(x, ...)
}
\arguments{
\item{x}{an object of class \code{guantile}.}

\item{...}{additional parameters (compatibility).}
}
\description{
Print a quantile object
}
\details{
Info
}
