% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print}
\alias{print}
\alias{print.evmm}
\alias{print.summary.ggpd}
\alias{print.quantile}
\alias{print.return_level}
\alias{print.VaR}
\alias{print.ES}
\alias{print.upper_bound}
\title{Printing Methods}
\usage{
\method{print}{evmm}(x, ...)

\method{print}{summary.ggpd}(x, ...)

\method{print}{quantile}(x, ...)

\method{print}{return_level}(x, ...)

\method{print}{VaR}(x, ...)

\method{print}{ES}(x, ...)

\method{print}{upper_bound}(x, ...)
}
\arguments{
\item{x}{an object created by \code{extrememix}.}

\item{...}{additional arguments for compatibility.}
}
\value{
A printed output of a model estimated with \code{extrememix}.
}
\description{
Collection of printing methods for various objects created by \code{extrememix}.
}
