% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.return_level}
\alias{print.return_level}
\title{Print a return_level object}
\usage{
\method{print}{return_level}(x, ...)
}
\arguments{
\item{x}{an object of class \code{return_level}.}

\item{...}{additional parameters (compatibility).}
}
\description{
Print a return_level object
}
\details{
Info
}
