% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.TVaR}
\alias{print.TVaR}
\title{Print a TVaR object}
\usage{
\method{print}{TVaR}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ES}.}

\item{...}{additional parameters (compatibility).}
}
\description{
Print a TVaR object
}
\details{
Info
}
