% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.upper_bound}
\alias{print.upper_bound}
\title{Print a upper bound}
\usage{
\method{print}{upper_bound}(x, ...)
}
\arguments{
\item{x}{an object of class \code{upper_bound}.}

\item{...}{additional parameters (compatibility).}
}
\description{
Print a upper bound
}
\details{
Info
}
