% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.VaR}
\alias{print.VaR}
\title{Print a VaR object}
\usage{
\method{print}{VaR}(x, ...)
}
\arguments{
\item{x}{an object of class \code{VaR}.}

\item{...}{additional parameters (compatibility).}
}
\description{
Print a VaR object
}
\details{
Info
}
