% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{quantile}
\alias{quantile}
\alias{quantile.ggpd}
\alias{quantile.mgpd}
\title{Quantiles Method}
\usage{
quantile(x, ...)

\method{quantile}{ggpd}(x, values = NULL, cred = 0.95, ...)

\method{quantile}{mgpd}(x, values = NULL, cred = 0.95, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}

\item{values}{vector of points where to estimate the predictive distribution}

\item{cred}{amplitude of the posterior credibility interval}
}
\description{
Compute and plot quantiles
}
