% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\docType{data}
\name{rainfall}
\alias{rainfall}
\title{Monthly Maxima Daily Rainfall in Madrid}
\format{
A  positive numeric vector of length 414. Observations where the monthly maxima are zero were discarded.
}
\source{
Instituto de Estadistica, Communidad de Madrid.
}
\usage{
data(rainfall)
}
\description{
Monthly maxima of the daily rainfall (measured in mms) recorded at the Retiro station in the city centre of Madrid, Spain, between 1985 and 2020.
}
\keyword{datasets}
