% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall_ggpd.R
\docType{data}
\name{rainfall_ggpd}
\alias{rainfall_ggpd}
\title{Rainfall FGGPD Output}
\format{
A list storing the output of the fggpd function over the rainfall dataset.
}
\usage{
data(rainfall_ggpd)
}
\description{
Estimated ggpd model over the rainfall dataset
}
\keyword{datasets}
