% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall_mgpd.R
\docType{data}
\name{rainfall_mgpd}
\alias{rainfall_mgpd}
\title{Rainfall FMGPD Output}
\format{
A list storing the output of the fmgpd function over the rainfall dataset.
}
\usage{
data(rainfall_mgpd)
}
\description{
Estimated mgpd model over the rainfall dataset
}
\keyword{datasets}
