% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_level.R
\name{return_level}
\alias{return_level}
\alias{return_level.evmm}
\title{Return Levels Method}
\usage{
return_level(x, ...)

\method{return_level}{evmm}(x, values = NULL, cred = 0.95, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}

\item{values}{vector of points where to estimate the predictive distribution}

\item{cred}{amplitude of the posterior credibility interval}
}
\description{
Compute and plot return levels
}
