% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.evmm}
\alias{summary.evmm}
\title{Summary Method}
\usage{
\method{summary}{evmm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{evmm}.}

\item{...}{additional parameters (compatibility).}
}
\value{
A printed summary of a model estimated with \code{extrememix} or any quantity associated with it.
}
\description{
Posterior estimates and credibility intervals for the parameters of extreme value mixture models.
}
