% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ggpd}
\alias{summary.ggpd}
\title{Summary of a ggpd}
\usage{
\method{summary}{ggpd}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ggpd}.}

\item{...}{additional parameters (compatibility).}
}
\description{
Summary of a ggpd
}
\details{
Info
}
