% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mgpd}
\alias{summary.mgpd}
\title{Summary of a mgpd}
\usage{
\method{summary}{mgpd}(object, ...)
}
\arguments{
\item{object}{an object of class \code{mgpd}.}

\item{...}{additional parameters (compatibility).}
}
\description{
Summary of a mgpd
}
\details{
Info
}
