% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR.R
\name{TVaR}
\alias{TVaR}
\alias{TVaR.ggpd}
\alias{TVaR.mgpd}
\title{TVaR Method}
\usage{
TVaR(x, ...)

\method{TVaR}{ggpd}(x, values = NULL, cred = 0.95, ...)

\method{TVaR}{mgpd}(x, values = NULL, cred = 0.95, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}

\item{values}{vector of points where to estimate the predictive distribution}

\item{cred}{amplitude of the posterior credibility interval}
}
\description{
Compute and plot TVaRs
}
