% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_bound.R
\name{upper_bound}
\alias{upper_bound}
\alias{upper_bound.evmm}
\title{Upper Bound Method}
\usage{
upper_bound(x, ...)

\method{upper_bound}{evmm}(x, cred = 0.95, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}

\item{cred}{amplitude of the posterior credibility interval}
}
\description{
Compute and plot return levels
}
