% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAIC.R
\name{WAIC}
\alias{WAIC}
\alias{WAIC.ggpd}
\title{DIC model selection criterion}
\usage{
WAIC(x, ...)

\method{WAIC}{ggpd}(x, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}
}
\value{
The WAIC of a model estimated with \code{extrememix}
}
\description{
Computation of the WAIC for an extreme value mixture model
}
