% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\alias{check_convergence.evmm}
\title{Convergence Assessment of MCMC Algorithms}
\usage{
check_convergence(x, ...)

\method{check_convergence}{evmm}(x, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extrememix}.}

\item{...}{additional arguments for compatibility.}
}
\description{
Plot of the traceplot and autocorrelation function for the 0.99 quantile from the posterior sample.
}
\examples{
\dontrun{
data(rainfall)
model1 <- fggpd(rainfall, it = 25000, burn = 5000, thin = 25)
check_convergence(model1)
}
}
