% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.evmm}
\alias{logLik.evmm}
\title{Log-likelihood Method}
\usage{
\method{logLik}{evmm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{evmm}.}

\item{...}{additional parameters for compatibility.}
}
\value{
The log-likelihood of a model estimated with \code{extrememix}
}
\description{
Computation of the log-likelihood of an extreme value mixture model (thus also \code{AIC} and \code{BIC} are available).
}
\examples{
logLik(rainfall_ggpd)
}
