% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.upper_bound}
\alias{plot.upper_bound}
\title{Plot Upper Bounds}
\usage{
\method{plot}{upper_bound}(x, xlim = c(min(x$bound), max(x$bound)), ...)
}
\arguments{
\item{x}{an object of class \code{upper_bound}.}

\item{xlim}{ciao}

\item{...}{additional parameters (compatibility).}
}
\description{
Plot Upper Bounds
}
\details{
Info
}
