% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{quant}
\alias{quant}
\alias{quant.ggpd}
\alias{quant.mgpd}
\title{Quantiles Method}
\usage{
quant(x, ...)

\method{quant}{ggpd}(x, values = NULL, cred = 0.95, ...)

\method{quant}{mgpd}(x, values = NULL, cred = 0.95, ...)
}
\arguments{
\item{x}{the output of a model estimated with \code{extremix}}

\item{...}{for compatibility}

\item{values}{vector of points where to estimate the predictive distribution}

\item{cred}{amplitude of the posterior credibility interval}
}
\description{
Compute and plot quantiles
}
